apps/python/pure/astra_db/requirements.txt
astropy
python-dotenv
unittest
